+++ src/mempool_args.cpp
@@ -7,13 +7,15 @@
 #include <kernel/mempool_limits.h>
 #include <kernel/mempool_options.h>
 
+#include <node/interface_ui.h>
 #include <util/system.h>
+#include <util/translation.h>
 
 using kernel::MemPoolLimits;
 using kernel::MemPoolOptions;
 
 namespace {
-void ApplyArgsManOptions(const ArgsManager& argsman, MemPoolLimits& mempool_limits)
+bool ApplyArgsManOptions(const ArgsManager& argsman, MemPoolLimits& mempool_limits)
 {
     mempool_limits.ancestor_count = argsman.GetIntArg("-limitancestorcount", mempool_limits.ancestor_count);
 
@@ -22,10 +24,12 @@ void ApplyArgsManOptions(const ArgsManager& argsman, MemPoolLimits& mempool_limi
     mempool_limits.descendant_count = argsman.GetIntArg("-limitdescendantcount", mempool_limits.descendant_count);
 
     if (auto vkb = argsman.GetIntArg("-limitdescendantsize")) mempool_limits.descendant_size_vbytes = *vkb * 1'000;
+
+    return true;
 }
 }
 
-void ApplyArgsManOptions(const ArgsManager& argsman, MemPoolOptions& mempool_opts)
+bool ApplyArgsManOptions(const ArgsManager& argsman, MemPoolOptions& mempool_opts)
 {
     mempool_opts.check_ratio = argsman.GetIntArg("-checkmempool", mempool_opts.check_ratio);
 
@@ -33,7 +37,19 @@ void ApplyArgsManOptions(const ArgsManager& argsman, MemPoolOptions& mempool_opt
 
     if (auto hours = argsman.GetIntArg("-mempoolexpiry")) mempool_opts.expiry = std::chrono::hours{*hours};
 
-    mempool_opts.full_rbf = argsman.GetBoolArg("-mempoolfullrbf", mempool_opts.full_rbf);
+    bool enable_replacement = argsman.GetBoolArg("-mempoolreplacement", /*fDefault=*/true);
+    if ((!enable_replacement) && argsman.IsArgSet("-mempoolreplacement")) {
+        // Minimal effort at forwards compatibility
+        std::string replacement_opt = argsman.GetArg("-mempoolreplacement", "");  // default is impossible
+        std::vector<std::string> replacement_modes = SplitString(replacement_opt, ",+");
+        enable_replacement = (std::find(replacement_modes.begin(), replacement_modes.end(), "fee") != replacement_modes.end());
+        if (enable_replacement) {
+            mempool_opts.full_rbf = (std::find(replacement_modes.begin(), replacement_modes.end(), "-optin") != replacement_modes.end());
+        }
+    }
+    if (!enable_replacement) {
+        return InitError(strprintf(_("-mempoolreplacement=%s is not supported by %s"), argsman.GetArg("-mempoolreplacement", "0"), PACKAGE_NAME));
+    }
 
-    ApplyArgsManOptions(argsman, mempool_opts.limits);
+    return ApplyArgsManOptions(argsman, mempool_opts.limits);
 }