+++ doc/release-notes.md
@@ -86,9 +86,9 @@ Changes to GUI or wallet related settings can be found in the GUI or Wallet sect
 New settings
 ------------
 
-- A new `mempoolfullrbf` option has been added, which enables the mempool to
-  accept transaction replacement without enforcing BIP125 replaceability
-  signaling. (#25353)
+- The `mempoolreplacement` option has been reintroduced to enable the mempool
+  accepting transaction replacement without enforcing BIP125 replaceability
+  signaling (when set to `fee,-optin`). (#25353)
 
 Tools and Utilities
 -------------------